+++ framework/views/alertDialog.js
@@ -32,6 +32,8 @@ limitations under the License.
       init: function(scope, element, attrs) {
         this._scope = scope;
         this._element = element;
+        this._attrs = attrs;
+
         this._element.css({
           display: 'none',
           zIndex: 20001
@@ -51,6 +53,8 @@ limitations under the License.
 
         this._deviceBackButtonHandler = $onsen.DeviceBackButtonHandler.create(this._element, this._onDeviceBackButton.bind(this));
         this._createMask(attrs.maskColor);
+
+        this._scope.$on('$destroy', this._destroy.bind(this));
       },
 
       /**
@@ -136,15 +140,19 @@ limitations under the License.
        * Destroy alert dialog.
        */
       destroy: function() {
+        this._scope.$destroy();
+      },
+
+      _destroy: function() {
+        this.emit('destroy');
+
         this._mask.off();
 
         this._element.remove();
         this._mask.remove();
         this._deviceBackButtonHandler.destroy();
 
-        this._scope.$destroy();
-
-        this._deviceBackButtonHandler = this._element = this._mask = null;
+        this._deviceBackButtonHandler = this._scope = this._attrs = this._element = this._mask = null;
       },
 
       /**