+++ framework/views/tabbarView.js
@@ -252,8 +252,6 @@ limitations under the License.
           if (this._tabItems[i] != selectedTabItem) {
             this._tabItems[i].setInactive();
           } else {
-            this._triggerActiveTabChanged(i, selectedTabItem);
-
             if (!needLoad) {
               this.emit('postchange', {index: index, tabItem: selectedTabItem});
             }
@@ -261,13 +259,6 @@ limitations under the License.
         }
 
         return true;
-      },
-
-      _triggerActiveTabChanged: function(index, tabItem){
-        this._scope.onActiveTabChanged({
-          $index: index,
-          $tabItem: tabItem
-        });
       },
 
       /**
@@ -388,11 +379,12 @@ limitations under the License.
        */
       _loadPageDOM: function(element, options) {
         options = options || {};
-        var pageScope = this._scope.$parent.$new();
+        var pageScope = this._scope.$new();
         var link = $compile(element);
 
         this._contentElement.append(element);
         var pageContent = link(pageScope);
+
         pageScope.$evalAsync();
 
         this._switchPage(pageContent, pageScope, options);
@@ -410,8 +402,19 @@ limitations under the License.
         this._switchPage(element, element.scope(), options);
       },
 
+      /**
+       * @param {Object} options
+       * @param {String} [options.animation]
+       * @return {TabbarAnimator}
+       */
+      _getAnimatorOption: function(options) {
+        var animationAttr = this._element.attr('animation') || 'default';
+
+        return TabbarView._animatorDict[options.animation || animationAttr] || TabbarView._animatorDict['default'];
+      },
+
       _destroy: function() {
-        this.emit('destroy', {tabbar: this});
+        this.emit('destroy');
 
         this._element = this._scope = this._attrs = null;
       }
@@ -436,7 +439,6 @@ limitations under the License.
 
       this._animatorDict[name] = Animator;
     };
-
 
     return TabbarView;
   });