+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 CHANGELOG
 ====
 
-v1.3.0-dev
+v1.3.0-beta
 ----
+ * ons-sliding-menu: Fixed [#544](https://github.com/OnsenUI/OnsenUI/issues/544).
  * ons-sliding-menu: Fixed [#464](https://github.com/OnsenUI/OnsenUI/issues/464).
  * ons-pull-hook: Fixed [#467](https://github.com/OnsenUI/OnsenUI/issues/467).
  * ons-pull-hook: Fixed issue where a error was thrown if the pull hook was destroyed before `$done()` was called.
@@ -14,8 +15,22 @@ v1.3.0-dev
  * ons-alert-dialog: Fixed [#511](https://github.com/OnsenUI/OnsenUI/issues/511).
  * ons-alert-dialog: Fixed [#512](https://github.com/OnsenUI/OnsenUI/issues/512).
  * core: `ons.createDialog`, `ons.createPopover` and `ons.createAlertDialog` can now be supplied with a scope object in order to specify the parent scope of the element. This makes data binding much easier.
+ * ons-loading-placeholder: Fixed [#541](https://github.com/OnsenUI/OnsenUI/issues/541).
+ * ons-loading-placeholder: The component can now be used anywhere.
  * ons-tabbar: Fixed [#530](https://github.com/OnsenUI/OnsenUI/issues/530).
  * ons-split-view: Fixed [#525](https://github.com/OnsenUI/OnsenUI/issues/525).
+ * ons-sliding-menu: Added event handler attributes.
+ * ons-alert-dialog: Added event handler attributes.
+ * ons-carousel: Added event handler attributes.
+ * ons-dialog: Added event handler attributes.
+ * ons-navigator: Added event handler attributes.
+ * ons-popover: Added event handler attributes.
+ * ons-pull-hook: Added event handler attributes.
+ * ons-split-view: Added event handler attributes.
+ * ons-tabbar: Added event handler attributes.
+ * ons-split-view: Fixed [#552](https://github.com/OnsenUI/OnsenUI/issues/552).
+ * ons-navigator: Added `animation` option to `popPage()`.
+ * ons-navigator: Added `replacePage()` method.
 
 v1.2.2
 ----